Fix spotlight being called multiple times
