+++ packages/rocketchat-ui-sidenav/client/sidebarHeader.html
@@ -12,7 +12,9 @@
 					</button>
 				{{/each}}
 			</div>
-			{{> toolbar}}
 		{{/with}}
+		{{#if showToolbar}}
+			{{> toolbar }}
+		{{/if}}
 	</header>
 </template>
