+++ packages/rocketchat-ui-sidenav/client/toolbar.html
@@ -8,7 +8,7 @@
 							<div class="rc-input__icon">
 								{{> icon block="rc-input__icon-svg" icon="magnifier"}}
 							</div>
-							<input type="text" class="rc-input__element rc-input__element--small" placeholder="{{getPlaceholder}}">
+							<input type="text" class="rc-input__element rc-input__element--small js-search" placeholder="{{getPlaceholder}}">
 							<div class="rc-input__icon rc-input__icon--right">
 								{{> icon block="rc-input__icon-svg" icon="plus"}}
 							</div>
