+++ packages/rocketchat-ui-sidenav/client/toolbar.js
@@ -1,49 +1,19 @@
+/* global menu, toolbarSearch */
 
-/* global menu */
 import { Meteor } from 'meteor/meteor';
-import { ReactiveVar } from 'meteor/reactive-var';
 import { Tracker } from 'meteor/tracker';
+import { ReactiveVar } from 'meteor/reactive-var';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 import { Session } from 'meteor/session';
 import { Template } from 'meteor/templating';
 import { TAPi18n } from 'meteor/tap:i18n';
 import _ from 'underscore';
 
-let isLoading;
 let filterText = '';
 let usernamesFromClient;
 let resultsFromClient;
 
-const selectorSearch = '.toolbar__search .rc-input__element';
-Meteor.startup(() => {
-	isLoading = new ReactiveVar(false);
-});
-
-const toolbarSearch = {
-	shortcut: false,
-	clear() {
-		const $inputMessage = $('.js-input-message');
-
-		if (0 === $inputMessage.length) {
-			return;
-		}
-
-		$inputMessage.focus();
-		$(selectorSearch).val('');
-
-		if (this.shortcut) {
-			menu.close();
-		}
-	},
-	focus(fromShortcut) {
-		menu.open();
-		$('.toolbar').css('display', 'block');
-		$(selectorSearch).focus();
-		this.shortcut = fromShortcut;
-	},
-};
-
-this.toolbarSearch = toolbarSearch;
+const isLoading = new ReactiveVar(false);
 
 const getFromServer = (cb, type) => {
 	isLoading.set(true);
@@ -118,14 +88,6 @@ Template.toolbar.helpers({
 		return placeholder;
 	},
 	popupConfig() {
-		const open = new ReactiveVar(false);
-
-		Tracker.autorun(() => {
-			if (open.get() === false) {
-				toolbarSearch.clear();
-			}
-		});
-
 		const config = {
 			cls: 'search-results-list',
 			collection: Meteor.userId() ? RocketChat.models.Subscriptions : RocketChat.models.Rooms,
@@ -137,7 +99,7 @@ Template.toolbar.helpers({
 			closeOnEsc: true,
 			blurOnSelectItem: true,
 			isLoading,
-			open,
+			open: Template.instance().open,
 			getFilter(collection, filter, cb) {
 				filterText = filter;
 
@@ -216,28 +178,17 @@ Template.toolbar.events({
 		return false;
 	},
 
+	'click [role="search"] input'() {
+		toolbarSearch.shortcut = false;
+	},
+
 	'keyup [role="search"] input'(e) {
 		if (e.which === 27) {
 			e.preventDefault();
 			e.stopPropagation();
 
 			toolbarSearch.clear();
-			$('.toolbar').css('display', 'none');
 		}
-	},
-
-	'click [role="search"] input'() {
-		toolbarSearch.shortcut = false;
-	},
-
-	'click .toolbar__icon-search--right'() {
-		toolbarSearch.clear();
-		$('.toolbar').css('display', 'none');
-	},
-
-	'blur [role="search"] input'() {
-		toolbarSearch.clear();
-		$('.toolbar').css('display', 'none');
 	},
 
 	'click [role="search"] button, touchend [role="search"] button'(e) {
@@ -249,4 +200,14 @@ Template.toolbar.events({
 			e.preventDefault();
 		}
 	},
+});
+
+Template.toolbar.onRendered(function() {
+	this.$('.js-search').select().focus();
+});
+
+Template.toolbar.onCreated(function() {
+	this.open = new ReactiveVar(true);
+
+	Tracker.autorun(() => !this.open.get() && toolbarSearch.close());
 });
