+++ packages/rocketchat_theme/client/imports/components/sidebar/toolbar.css
@@ -2,8 +2,6 @@
 	position: absolute;
 	left: 10px;
 
-	display: none;
-
 	width: 100%;
 	margin: 0 -10px;
 	padding: 0 calc(var(--sidebar-default-padding) + 10px);
